critera.md: remove "code coverage" requirements.
Code "coverage" is really just "test coverage" and is not a good metric for any sort of "assurance" other than someone wrote a bunch of tests.  See http://martinfowler.com/bliki/TestCoverage.html for more details as to why I wouldn't even put this on a list of something to try to achieve toward.