Updating OSS Project Basics
First attempt at reformatting and rewording some of the criteria in the OSS Project Basics.  The intent is to try and make the criteria easy to identify by placing them on separate lines or by breaking down lists.  I've also separated information about testing, future releases and rationale from the criteria.

I haven't changed too much of the content.  Let me know if this type of change is something the project is interested in - if so I'll continue through the rest of the Basic Criteria.