Add various images from Karol Munoz
Karol Munoz created, as part of this project, some
PSD files and their generated .png files named:
    home_badge@2x.png            result_symbol_dash@2x.png
    home_badge.png               result_symbol_dash.png
    questions_page_badge@2x.png  result_symbol_question@2x.png
    questions_page_badge.png     result_symbol_question.png
    result_symbol_check@2x.png   result_symbol_x@2x.png
    result_symbol_check.png      result_symbol_x.png

They may be used under either the MIT license or the
Creative Commons - Attribution version 3.0 or later (CC-BY-3.0+) licenses.
For license details, see:
https://opensource.org/licenses/MIT and
https://creativecommons.org/licenses/by/3.0/us/
This license was confirmed on 2016-04-11 in the email from her
with subject line 'Re: Images for green check (yes), X (no), "?" (unknown),
yellow-dash (passing, but not fully met)?'.

We intend to use these result symbols instead of simply
Thumbs up/Thumbs down, but it's useful to have the images before
we write the code (so we can use the images to test the code).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>