Enable hiding of N/A along with met criteria
Enable hiding of N/A along with met criteria.
The hiding is intended to help people see what is left to do for a badge,
and this will help people examine what's left.
We will continue to display the "barely met" criteria, because we
want to encourage people to do better, but in the case of N/A status
there's nothing left that needs doing.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>