Add GitHub oauth tests
This is unusually complicated automated testing.

The test added runs in two modes. When the vcr cassette is missing and a password is supplied
as an environment variable, it connects to Github and saves all of the back and forth into a cassette.
When the cassette is there, it mocks the OAuth connection and uses the recorded HTTP transactions
to test our parsing of Github info. This causes our code coverage to jump from 94.06% to 96.52%.
The same approach can be used with detectives.

Nothing security-relevant is exposed in the repository.
The password is entered as an environment variable by
test/features/github_login_test.rb.
We log out at the end to invalidate the OAuth token that Github gives us.
The values stored in .env are not the production (actual secret) values.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>
Signed-off-by: Dan Kohn <dan@dankohn.com>