Record various project statistics daily. Fixes #331
Add a mechanism to record information about projects
in the new table project_stats.
It does include a primitive display mechanism,
but it is temporary.  The purpose of this commit is
primarily to record information.
We intend to later add a real viewing mechanism
once there's something to view.

The new "rake daily" task adds the current values
of the various statistics to the new project_stats table.

The original code was by David A. Wheeler,
with a number of changes by Dan Kohn.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>
Signed-off-by: Dan Kohn <dan@dankohn.com>