All repo and project URLs with UTF-8 encoding. Fixes #402
Modify the handling of repo and project URLs to allow
embedded UTF-8 encoding, but with careful limitations to protect security.
In particular, we are limiting URLs so the BadgeApp has less utility for
attacking other sites (e.g., we don't allow "?"), but this allows
%-encoding of a number of characters.  This change allows %20 as the
encoding for space.  It requires that URLs, once decoded, be valid UTF-8.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>