Panes change URL, and jumping to criteria shows pane. Fixes #218
This change means that selecting a pane will change the URL fragment,
*and* if you request a specific criteria id, the displayed pane will
change to show it.

An unexpected problem: At least on Firefox, on initial load it
goes to the correct pane on initial load, but the ScrollIntoView
doesn't really move to the intuitively correct location.
Similarly, selecting a pane jumps down to the first question, and
doesn't show the larger pane (or the other tabs).
Improvements welcome.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>