Revert local password to simply 7+ characters
Revert local password rules back to "7+ characters".
Implementing them as validation rules means that users
can't log in using their previous passwords if they
fail the new rules.  Stronger password
checks will need to be implemetned in a different and
more backwards-compatible way.  This change *keeps* the
loading of a "bad password" list, since we *do* intend to
use it eventually - just in a better way.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>