Tweak sites_https: CloudFlare CDN is okay for GitHub pages + custom domains
Many projects use GitHub pages with custom domains.
GitHub pages support HTTPS, but only if there are no custom domains.
GitHub does not directly support HTTPS for custom domains, but there is
a workaround involving using a CDN (such as CloudFlare) as a proxy.
Make it clear that this workaround is fine.  While we're at it,
give guidance for people who use GitHub pages without custom domains,
and make it clear that supporting HTTP is *permitted*
(we just require that they support HTTPS).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>