Modify rationale re: SemVer - in some cases it's not so helpful
There is some controversy about SemVer, esp. when there's only a
single running instance (via continuous delivery).
It is only a SUGGESTED, but it may not be obvious to readers
WHY it is only SUGGESTED, so let's make that clear.
This should help them decide if it's appropriate to their circumstance.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>