Exclude old db migrations from Rubocop (prep for rubocop 0.47.1)
This excludes older database migrations from Rubocop checking.
Migrations, once written, normally never change, since they are
used to modify the database.  So it's a waste of time to check
ancient migrations.  Even more importantly, some migrations
trigger many warnings in a newer version of rubocop (version 0.47.1).

In particular, db/migrate/20151005140708_default_status.rb
triggers 63 "ReversibleMigration" errors (a large number!), because it
has a number of "change_column_default" statements without :from and :to.
We are never going to reverse that (it's too far back in time), so
its irreversibility is irrelevant.

Newer migrations continue to be checked, so we can quickly
detect problems in more recent migrations.