Require markdown lists to be surrounded by blank lines, and fix
This adds rule MD032 to markdownlint, which requires that
lists be surrounded by blank lines within markdown files.

This detects a lot of malformed text that should be in a list but is not.
This is also important for portability: Some markdown formatters
require blank lines around lists, while others don't.
To see this portability problem in action, see:
http://johnmacfarlane.net/babelmark2/?text=Is+this+a+list%3F%0A*+First+bullet%0A*+Second+bullet%0A

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>