Update gem rainbow (you may need to run "gem update --system")
This updates the gem rainbow from 2.1.0 to 2.2.0.

Previous "bundle install" efforts failed in some situations
with the message "can't modify frozen String".
As a quick workaround we forced use of the old version
until we worked out a better fix.

However, on investigation, the root cause turns out to be
a need to upgrade rubygems, using:
  gem update --system

More information:
https://github.com/bundler/bundler/issues/5357
https://github.com/sickill/rainbow/issues/44

So we can now update the gem rainbow.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>