Indicate badge entry data is licensed as CC-BY-3.0+ from now on
Clearly indicate that all badge entries from now on are CC-BY-3.0+.
Copyright would probably be limited on badge entry data anyway,
but we want to encourage people to see and use this data, so having
a clear common license for the badge entry data is helpful.
We could have chosen CC0 (essentially public domain), which would be
easier for later users because no attribution would be needed.
However, attribution might encourage people to provide data
(and incentivize them to enter *correct* data).
All other non-code information in the badge project is also CC-BY-3.0+,
so this is consistent.

We only indicate this license if there's an updated_at after a date;
this is so the license is only shown if the user has accepted it
(as indicated selecting save or submit just under text declaring this).
We can't just relicense other people's data without their permission;
this seemed like the simplest way to deal with that.
We could go require more ceremony (by requiring a checked checkbox),
but we're trying to maximize ease-of-use here,
and this is essentially the same way that Wikipedia accepts changes,

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>