Make legal statements much more prominent
This change adds more direct links to the terms of use.

Previously the legal "terms of use" were findable but it took
two clicks.  We previously noted that this was a CII project and
referred to the CII main page.  Anyone who wanted to find the
terms of use should have realized that since there was nothing
else stated, they should go to the overall project, and from there
they could easily find the "terms of use" link which pointed to the
Linux Foundation "terms of use" page at:
  https://www.linuxfoundation.org/terms

We now improve this situation.
With this change, the terms of use are directly and
prominently on the front page.  Links to the terms of use and
privacy policy are also now included in the footer of every page.

The "terms of use" page is important for a variety of reasons.
The page made it clear that submitting material in violation of
copyright or confidentiality agreements was *not* acceptable,
and that users would have their accounts terminated if
LF determined that they were repeat violators.
It also provided information on the Linux Foundation
DMCA designated agent.  These are required for the US law
DMCA "safe harbor" as explained here:
http://www.natlawreview.com/article/
website-operators-must-comply-new-dmca-agent-registration-procedures-or-risk
This page also makes it clear that contributed materials are
under the CC-BY-3.0 license unless otherwise noted.

While we're at it, we also make it clearer that CII is itself a
Linux Foundation project.  At the
Open Source Leadership Summit 2017 in Lake Tahoe, California,
the Eclipse Foundation made a repeated point that projects should
ensure that they note their supporting foundations - it's good
for both the foundation and the supporting project.
The website *already* had a "Linux Foundation" logo
on every page footer, but here we also add "Linux Foundation" to
the text on the front page, making it even more obvious.

We also add hyperlinks on the front page to
both the LF and CII websites, to make it easier for people to
go to those sites for more information.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>