More clearly state that all content is at least CC-BY-3.0
Make it more obvious that all materials are at least licensed
under the Creative Commons Attribution version 3.0 license (CC-BY-3.0).
Our site is clearly a CII page (it says so!), and the
the CII main page <https://www.coreinfrastructure.org/>
points to the overall Linux Foundation "terms of use" at
<https://www.linuxfoundation.org/terms>.
Here are a few quotes from the "terms of use" that make this clear:

"The Linux Foundation and its websites provide a framework for
discussion and user generated information to expand the knowledge base
of Linux-related information....
Terms and Conditions of Use for the Linux Foundation's Web Sites
BY ACCESSING, BROWSING OR USING THIS WEB SITE, YOU ACKNOWLEDGE THAT YOU
HAVE READ, UNDERSTAND AND AGREE TO BE BOUND BY THESE TERMS AND CONDITIONS...
Except as otherwise provided, content on this site, including all
materials posted by the The Linux Foundation, is licensed under a Creative
Commons Attribution 3.0 License...
Some Linux Foundation Workgroups or other projects that may be hosted by
the Linux Foundation may have license terms or Contributor Agreements
that are specific to the Workgroup or project...
to the extent there is a conflict, those license
terms or agreements take precedence over these Terms of Use...
All Other Submissions. User Content that is not in the form of source or
object code, including but not limited to white papers, dissertations,
articles or other literary works, power point presentations,
encyclopedias, anthologies, wikis, blogs, diagrams, drawings, sketches,
photos or other images, audio content, video content and audiovisual
materials, will be governed by the Creative Commons Attribution 3.0."

It should be painfully obvious to anyone dealing with the best practices
website that it's a CII site, and that the CII terms of use apply.
The best practices website clearly identifies itself as a CII (and LF) site.
The main site, bestpractices.coreinfrastructure.org, is even a
subdomain of "coreinfrastructure.org".  The CII, in turn, points to
these clearly-stated terms of use.
Therefore, anyone who cared to check could have quickly determined that
submissions were released under at least CC-BY-3.0.

That said, we got at least one question at the LF Summit 2017 in Lake Tahoe,
and it's good to make licensing information as obvious as we can make it.
We also decided to add "or later" for newer contributions, to maximize
potential data use by others.
We expressly tell people that newer materials are CC-BY-3.0+
("3.0 or later"), to maximize compatibility, but we can't relicense
older materials.

So in the case of older materials, this commit makes
it much more obvious what the license is for everything:
for older materials, for newer materials, and for the entire set.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>