Cache showing entire project
Cache an entire project entry when viewed read-only.

Note that we do NOT cache editing the entire project, because
in almost all cases an edit means we haven't edited *exactly*
the same version before (so it'd be pointless to cache).

This is in addition to caching the status chooser. If changes
have happened to a project entry, the entire-project cache
entry will fail, but many of the status chooser entries will
still produce cached values (if they've been seen before).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>