Cache project_stats page
This greatly speeds up redisplay of the /project_stats page, e.g.,
on my local system it goes from 4415ms to 185ms.  The cache is keyed
off the number of project_stats entries, so cached entry will stop
being used when a new project_stat is created (which is daily).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>