Change application.css to application.scss so @import is optimized
Rename "application.css" to "application.scss".  This means that the
Sass CSS preprocessor will run on the file first to generate the
final "application.css" file, and in the process, it optimizes out
the @import statements so that there is a single file
(so the browser doesn't need to separately request them).
This reduces latency.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>