Cache almost entire project entry when showing it (is_disabled==true)
Cache almost the entire project entry when showing it.
We don't cache the entry on edit, because it's about to change
(and is thus pointless).

Local testing shows an 80% increase in speed in the server-side
completion time when displaying project entries for read (after the
first display).

We could cache smaller fragments of the project entry for editing,
but it's not clear it's worth it (at this time).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>