Add caches to atom feed, fixes #702
Feed users will poll the /feed URL, so we want to respond quickly
when there's no change.  Add caches to make the atom feed faster.
This caches the whole result (in the case where there's been no change),
as well as individual entries (so that if there's been a change, we can
still use cached results in many cases).

This cache is not used in during tests. There appears to be
a bug in the test framework's handling of caches in this case,
and disabling the cache during testing works around the problem.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>