Cache project data in JSON format
This provides a mild speed-up when querying project data in JSON format.
On my local system, it goes from 26.6ms to 21.2ms, a speedup of 1.25,
when the same data is queried more than once.

The cache key doesn't include the project owner's name, because
changing names is an extremely rare operation.  If a name change
happens after the project JSON data has been cached, *and* the cache
is used, then there will be a window of time where the old name is used.
This will probably never occur in practice, and it's not wrong,
it's just an older name (the project data itself will be up-to-date).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>