Merge project JSON builder into 1 file (DRY it)
Move the JSON format generator for a project into a single
file, the partial _project.json.jbuilder. Then call the new
partial from "index" and "show".  This helps keep the code
DRY (don't repeat yourself).

This improves maintenance, though at a small performance cost:
index is slightly slower.  But index for JSON is rare, so that's
not a big deal.  "Show" might be slower too, but it's cached and
is quite fast anyway, so that shouldn't matter.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>