Add test: project show in JSON format (regression test)
Add a test to show project data in JSON format.

Remarkably, we somehow didn't have such a test, and had a bug
in showing project data in JSON format.  The bug is now fixed.
This is a regression test to ensure similar problems
are automatically detected by the test suite in the future.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>