Omit long unused fields from criteria.js.erb
Omit long unused fields from criteria.js.erb, instead of just
sending the entire content of the criteria.yml file in JSON format.

This slightly speeds first-page view, because this file
(once processed) is sent to clients when viewing content.
This is basically irrelevant for those with good network bandwidth,
but it should help people with poor network bandwidth.
It only affects first-page view, because once
this file is downloaded it's cached by normal web browers.

This only helps a little bit and only in specific cases,
but it's an easy and obvious optimization to make, so let's do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>