Project_stats: Delete useless cache entry & remove HTML table
The /project_stats page is big, and its cache entry quickly uses
up space for other things.  Delete the cache entry for the obsolete
dataset (if it's around), since it's no longer useful.
Also, remove the hidden HTML table; it takes up a lot of space
and doesn't provide any new data (if you wan that data, the JSON
file can provide the same thing in an easily-processed form).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>