Modify /feed to select only fields to be displayed
The /feed only displays a small set of the project fields, so only
extract the ones we use.  This optimization is worth it because
users poll the feed (so we see it often) *and* it can include many projects,
and projects have a large number of fields.

This somewhat reduces execution time. In local testing this
reduces uncached feed creation from 350ms -> 310ms.
It also reduces memory pressure since the temporary objects we create
will only include the fields we'll later use.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>