Performance improvements: /feed and don't recalc for badge_level
Improve performance.  In /feed, only get the project fields we need
(the previous patch didn't quite do this, due to a known bug in Rails).
This only works if we don't constantly recalculate everything from
scratch when determining the badge_level (since that would required
the full dataset), but we were accidentally recalculating
everything when determining the badge level instead of
simply using the badge_percentage.  That was slow but worked when we
always had every data field, but this change removes that data.
As a result, we must modify the code so it
simply uses the already-calculated badge_percentage instead of
recalculating everything.

Sample performance for empty /feed goes from 350ms (original) to 211ms.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>