Modify "delete user" so projects always have an owner
Project entries should always have an owner, as an invariant.
This tweaks the "delete user" code so this invariant is maintained.
It does this by transferring ownership of projects to the
deleting admin from the user being deleted, and by preventing
admins from deleting themselves (this would otherwise be a loophole).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>