Simplify logic in _status_chooser; improve JavaScript Consistency
We added the routine get_criterion_status in project.rb in order to
retrieve a symbol for a particular criterion's status these are:

* 'criterion_passing',
* 'criterion_barely'
* 'criterion_url_required'
* 'criterion_justification_required'
* 'criterion_unknown'
* 'criterion_failing'

This allows for much simpler login in the _status_chooser view.

We also updated the project-form.js so that it produces results consistent
with _status_chooser.  In doing so we also fixed login_test.

The next step is to refactor the JavaScript so that the logic is almost
identical to that used in the ruby code.  This will make our code base
much easier to maintain.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>