Add separate /reminders URL for details about reminders
Add a new separate /reminders URL for details about reminders, and
remove similar information from "/project_stats".

One reason to do this is to separate views with different
security properties into different pages.

The /project_stats page is visible to everyone (even if people
who are not logged in), and it *must* allow embedded scripts because
that's how the graph tool works (so we have to include a CSP exception).
If only dates and numbers can be included, the risks are lower.

Here we move the /reminders URL into its own page.
It can only be viewed by logged-in admins, because it includes
email addresses.  Putting it in a separate page means that we can
block viewing directly in the controller, and we can also have a
stronger CSP (so if the HTML escaping fails, CSP will still protect
the admins - who we *definitely* want to protect).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>