Clarify criteria: *sites* or project results, project definitions
This is in response to:
https://github.com/linuxfoundation/cii-best-practices-badge/issues/270

Some criteria were not as clear as they needed to be as to whether
they applied to the software produced by the project, or by project sites.

Document much clearer definitions for terms related to projects.
Then based on those terms, make the criteria clearer and consistent.
In particular, the crypto criteria really focused on the software
produced by the project (as indicated by the big button), but this
wasn't so obvious in the text.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>