Optimize JavaScript; Shrink project form by removing duplicate text;
This changes the javascript so that functionality that is only relevant
to editing a project form is only loaded when editing the project.
Some calls to the jquery function parents() were changed to closest()
as closest() is faster and actually what we wanted.

Additionally when the project is not being edited, we no longer include
a hidden text box with the justification data for evaluating criterion status.
We instead read the generated markdown field.  This shrinks the project
document when not editing.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>