Upgrade rubocop to 0.48.1; Closes #577; Upgrade pronto-rubocop to 0.8.1
This commit makes all the neccessary style changes to comply with the
ruby style guide as of rubocop 0.48.1. In addition, we enforce
Style/NegatedIf only for the postfix case as is now allowed by rubocop.

There is currently an issue with the Lint/AmbiguousBlockAssociation
cop returning erroneous warnings in app/models/project.rb that has been
fixed upstream in rubocop but not yet released.  For now we have
disabled this rule where needed, but need to re-enable it on the next
release.  This has been noted in the code with a TODO.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>