More Javascript Optimizaitons;
We had previously had to repeatedly query the document when setting
a panel's satisfaction level or changing the progress bar.  This lead to
a lot of repeated function calls.  Now we store the current state of the
criteria in a hash that we can change and access at any time.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>