Add met_justification_required to static_analysis and neccessary tests.
This adds the requirement that a justification be entered for the
static_analysis criterion.  This will not affect the status of any
currently passing projects unless they edit.  If a project has selected met
and edits their entry, they will be flashed a message to add a justification
for static_analysis. Only after they save will any absense of a justification
for static_analysis affect their badge level.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>