Move criteria.yml for better asset pipeline management; preserve its history
app/assets/javascripts/criteria.js.erb depends on critveria.yml, but
the asset pipeline had no way of knowing if criteria.yml changed.
So if we added or changed criteria, criteria.js was not being
recognized as having a resulting changed leaving us issues.
This attempts to resolve this by adding a dependency tag in criteria.js.erb
and moving criteria.yml into the application path.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>