Add sites_password_security, fixes #728
There are many other ways to harden a project site, but this
seems like the most basic measure and one that is widely agreed on.
So, add this as a key criterion for hardening sites.

At some point this should probably be moved down from passing+1 to passing;
perhaps this should be in "future".