Configure to store individual project json files on Fastly.
We have been seeing a large number of json requests for some particular
projects' json files coming from a single IP (1 request ever 30 seconds).
While it is not a problem at this time, it could become a problem down
the line should more people do it.  This mitigates the possible issue by
allowing the project json files to be cached on Fastly.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>