Fix some capitalization errors in home page locale 'fr' (French)
File home.fr.html.erb is the static page for the home page in the
'fr' (French) locale.  This was mainly created using Google Translate,
but that has the odd habit of capitalizing the first letter of each
new line.  Fix that.

The translation here is still poor; it's just a machine translation
for the most part (I did the "free software" translation).
Still, it's a decent stub so that we can see if the internationalization
works.