CONTRIBUTION.md: Refer to ActiveSupport::HashWithIndifferentAccess
Rails 5.1 has a "soft deprecation" of references to the top-level type
HashWithIndifferentAccess; its documentation recomments refer to
ActiveSupport::HashWithIndifferentAccess instead.
We don't directly refer to HashWithIndifferentAccess except in documentation,
but let's include the recommended name in the documentation so that
if we ever *do* use it we'll use the correct one.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>