Increase rubocop's permitted Metrics/AbcSize to support internationalization
The default Metrics/AbcSize is too small, and
often complains about code that is easy to understand and easy to test.

This is exacerbated by the internationalization work.
AbcSize is increased by every method call, and in Rails, text
marked for internationalization creates another call.
Thus, an internationalized program in Rails will typically have methods
with a larger AbcSize than a program that is not internationalized.

This change slightly increases the maximum permitted value to
reduce the noise.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>