Internationalize flash messages from sessions
Internationalize flash messages based on sessions.
The trick here is that "new.html.erb" has to use URLs
in the "?locale=" form, not the "/:locale", to interact properly
with other libraries.

The interactions in a session can be subtle (due to callbacks and
redirects), so I've intentionally squashed all
internationalization strings into one level below "sessions".
This way nobody has to worry about using the "wrong" method.