Internationalize "passing" and "in_progress" emails
Modify the email mechanism so the views are internationalized.

This isn't complete, because there's no mechanism for setting the
appropriate locale when sending the messages - so this also includes
a TODO comment to make that more obvious.  However, it's worth doing this
now, so that translators will have the text and keys needed to do the
translation.