Mark project descriptions' language (currently en)
When showing a list of project descriptions, the text of the
project description need not be the same as the user's current locale.
Thus, we need to mark the description text separately.
Currently we'll just mark them as the constant 'en'.
If we support adding text in other languages that will need to be more
complicated, but this at least makes it much easier to reliably use
machine translation programs (because they will now have accurate marking
of the locale in use).