Email user when their user data changes
This alerts the user whenever the user data changes -
this is useful for security, since if an attacker gains control
over their account the user will be alerted about any permanent
change in the account status.

This doesn't prevent a user from changing his email address; if the
user changes the address to someone else, then the "someone else"
has control over the user's account (which is, after all, what the
user requested).