Add field preferred_locale to User record
This is a small change, but there was extra debugging time because
"validates ... inclusion: LIST" requires that LIST be a sequence of
*strings*, but I18n.available_locales is actually a sequence of *symbols*.
When inclusion is provided a sequence of symbols, none of the strings
match (unsurprisingly).  This produces a long set of somewhat-mysterious
error messages of this form:
ActiveRecord::RecordInvalid: Validation failed:
Preferred locale is not included in the list
I record this information for "next time".
This is another case where "strings and symbols can be used
interchangeably, except when they can't".