When user logs in, switch to the user's preferred locale
When the user logs in, modify the forwarding URL so that we switch
to the user's preferred locale, then use that modified URL.
The user-visible effect is that logging in causes the user
to switch to the user's preferred locale.

This change has a lengthy set of tests, because there are a
number of special cases & we want to reduce the risk of mishandling one.
File test/helpers/sessions_helper_test.rb includes a large number
of different cases (to reduce the risk of missing a key case).
File test/features/login_test.rb ensures that users switch when they
log in; this part can't be a unit test, because in unit testing we
optimize login to speed test execution. This optimization skips
steps that are normally irrelevant, but they matter
when testing this feature.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>