Fix some assert_equal (expected value is supposed to be first)
Change some assert_equal statements so the expected value is first.
When tests pass the order doesn't matter, but when tests fail the
output can be confusing (because the output presumes the expected
value is first).  This change is useful so that any future test failures
will produce clearer output, *and* will increase the likelihood that
future changes will use the "expected value is first" convention.