Update project.rb and associated views for higher levels.
This updates a number of files to be compatible with the new criteria structure
there is still work to be done, but we can now render individual projects, for
the passing level and also edit them.  We still need to update the mailers,
tests, project_stats, and some of the static pages.  We also need to insert
the the placeholders into the criteria.js

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>