Switch locale on login only current locale in :en
On login we switch to users's preferred locale - but only switch
if the current locale is :en.  Any other locale is an intentional
choice, so in those cases we should retain what the user has already
specifically selected.  This should eliminate some user confusion;
the default preferred_locale in 'en', but if a user hasn't changed that
*and* is already in another locale, the user will simply stay in the
locale they've separately switched to.