Move home page text into config/locales/*
Move the home page text into the config/locales/* files so that
translation of its text can happen the same way as other translations.
One of the original reasons this was handled differently was for
speed (the home page gets hit often); we now cache the home page
to achieve similar performance.