Fix flapping test with phantomjs 2.1.1; rearrange sessions/new slightly
On phantomjs 2.1.1, one fo the tests in features/login_test.rb would flap.
The flapping would not occur on some earlier versions of phantomjs.
This commit moves the code which tests if logged in users are redirected to
the root url when trying to get /login to
test/controlers/sessions_controller.rb.
This is a better location for the test anyway, also but has the added benefit
of preventing the flapping test.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>