Update project-form.js for locales and higher levels
This updates project-form.js to take in the newly structured
CRITERIA_HASH_FULL and use it.  It is now able to read the
locale from the url using the getLocale function.

We have also inserted a dummy function getLevel() which
for now returns '0', but can be modified once we have decided
how levels will be encoded into the URL.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>