On form submission stay on current level
Previously we modified "update" so that it would accept a
criteria_level, and when provided, it would redirect to that criteria_level
after saving.

This commit modifies the form submission to *provide* this data to
update.  The result: When users save their project data, they stay
at the same criteria_level.