Cleanup: Enable auto-generation of all status choosers in a minor area
This change should simplify the view templates, as well as
criteria maintenance, because it lets us replace a sequence
of "render_status" calls with a single call to render_minor_status.
Note: this *does* mean that if the name of a major or minor group key
changes, you have to change the call to match.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>