Add project headings to locale (so they can be translated)
We need the project headings to be in config/locales/en.yml
so they can be translated to other languages.
This commit adds the project entry headings to make this possible.

We've chosen to put them all in the top-level key "headings";
that way duplicates are automatically handled.
The keys are the actual heading text (like gettext), to make it
easy to access by the program.

This commit involved some cleanup of heading upper/lowercase, since it
will be more trouble to change them later.

This list was found by first running this in "rails console":
CriteriaHash.each { |level|
 level[1].each { |c|
  data = c[1]
  puts '"' + data['major'] + '" : "' + data['major'] + '"'
  puts '"' + data['minor'] + '" : "' + data['minor'] + '"'
 }
}

The output was run through "sort -u", then through
YAML lint <http://www.yamllint.com/>, and then fixed up by hand.

From here on, the expectation is that headings will only be
added or changed infrequently, and they'll just be modified
directly.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>