Remove time_ago_in_words from projects/show
The rails_best_practices tool warns that we shouldn't be
using method time_ago_in_words, in part because it doesn't work
with caching:
https://rails-bestpractices.com/posts/2012/02/10/not-use-time_ago_in_words/

This is correct - we cache, and the result of time_ago_in_words
will only be correct when we generate it in the first place
(not in later uses from the cache).

This commit removes the use of time_ago_in_words; we still display
datetimes, so the key data remains.  In the future we could choose
to add client-side JavaScript to do the same thing.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>